Struct nphysics::resolution::VelocityConstraint

pub struct VelocityConstraint {
    pub normal: Vect,
    pub weighted_normal1: Vect,
    pub weighted_normal2: Vect,
    pub rot_axis1: Orientation,
    pub weighted_rot_axis1: Orientation,
    pub rot_axis2: Orientation,
    pub weighted_rot_axis2: Orientation,
    pub inv_projected_mass: Scalar,
    pub impulse: Scalar,
    pub lobound: Scalar,
    pub hibound: Scalar,
    pub objective: Scalar,
    pub id1: isize,
    pub id2: isize,
    pub friction_limit_id: usize,
    pub friction_coeff: Scalar,
}

A constraint of velocity at a point of contact.

Fields

normal

The contact normal.

weighted_normal1

The contact normal multiplied by the first body's inverse mass.

weighted_normal2

The contact normal multiplied by the second body's inverse mass.

rot_axis1

The first body rotation axis.

weighted_rot_axis1

The first body rotation axis multiplied by its inverse inertia.

rot_axis2

The second body rotation axis.

weighted_rot_axis2

The second body rotation axis multiplied by its inverse inertia.

inv_projected_mass

The inverse of the sum of linear and angular inertia of both bodies.

impulse

The total impulse applied.

lobound

The lower bound of the impulse.

hibound

The upper bound of the impulse.

objective

The target delta velocity.

id1

The id of the first body.

id2

The id of the second body.

friction_limit_id

The id of the friction constraint.

friction_coeff

The friction coefficient on this contact.

Methods

impl VelocityConstraint

fn new() -> VelocityConstraint

Creates a new velocity constraint with all terms initialized to zero.